Added toolbar. And applied full-transparent status bar
