+++ core/src/main/java/in/koreatech/koin/core/activity/ActivityBase.java
@@ -2,7 +2,9 @@ package in.koreatech.koin.core.activity;
 
 import android.content.Context;
 import android.content.pm.ActivityInfo;
+import android.graphics.Color;
 import android.os.Bundle;
+import android.view.WindowManager;
 
 import androidx.annotation.Nullable;
 import androidx.annotation.StringRes;
@@ -10,6 +12,7 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import in.koreatech.koin.core.progressdialog.CustomProgressDialog;
 import in.koreatech.koin.core.progressdialog.IProgressDialog;
+import in.koreatech.koin.core.utils.StatusBarUtil;
 
 
 public class ActivityBase extends AppCompatActivity implements IProgressDialog {
@@ -20,6 +23,7 @@ public class ActivityBase extends AppCompatActivity implements IProgressDialog {
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         context = this;
+        StatusBarUtil.makeStatusBarFillTransparent(getWindow());
         try {
             setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_PORTRAIT);
         } catch (IllegalStateException ignore) {
