+++ koin/src/main/AndroidManifest.xml
@@ -24,7 +24,9 @@
         android:theme="@style/AppTheme"
         android:usesCleartextTraffic="true"
         tools:replace="android:name">
-
+        <activity android:name=".ui.main.MainActivity"
+            android:label="@string/app_name"
+            android:theme="@style/KAPTheme.NoActionBar" />
 
         <receiver android:name=".ui.dining.DiningWidget">
             <intent-filter>
@@ -93,7 +95,7 @@
             android:name=".ui.forgotpassword.ForgotPasswordActivity"
             android:theme="@style/KAPTheme.NoActionBar" />
         <activity
-            android:name=".ui.main.MainActivity"
+            android:name=".ui.main.MainActivityOld"
             android:label="@string/app_name"
             android:theme="@style/KAPTheme.NoActionBar" />
         <activity
