+++ koin/src/main/java/in/koreatech/koin/ui/splash/SplashActivity.java
@@ -18,6 +18,7 @@ import com.crashlytics.android.Crashlytics;
 import in.koreatech.koin.BuildConfig;
 import in.koreatech.koin.R;
 import in.koreatech.koin.core.activity.ActivityBase;
+import in.koreatech.koin.ui.main.MainActivity;
 import in.koreatech.koin.ui.splash.presenter.SplashContract;
 import in.koreatech.koin.data.sharedpreference.UserInfoSharedPreferencesHelper;
 import in.koreatech.koin.core.network.RetrofitManager;
@@ -26,7 +27,7 @@ import in.koreatech.koin.util.FirebasePerformanceUtil;
 import in.koreatech.koin.core.toast.ToastUtil;
 import in.koreatech.koin.ui.splash.presenter.SplashPresenter;
 import in.koreatech.koin.ui.login.LoginActivity;
-import in.koreatech.koin.ui.main.MainActivity;
+import in.koreatech.koin.ui.main.MainActivityOld;
 import io.fabric.sdk.android.Fabric;
 
 public class SplashActivity extends ActivityBase implements SplashContract.View, VersionDialogClickListener {
